version bump!
We now have a single program rather than running different problems through examples.
The program has speciation enabled by default, and allows for loading parameters from a config file. Saves results within the `data/<problem>/` directory, including the config file and the results.json.

Still has the issue of non-reproducibility. I believe this may be because the test generation isn't based on this seed, but I'm unsure. To be investigated.